Split out DocumentWriterExtensions class
